+++ src/pages/Admin/Home/Dashboard/AdminTaskList/ListBody/container.tsx
@@ -1,5 +1,5 @@
 import React, { Fragment, FunctionComponent } from "react";
-import { List } from "./component";
+import { ListBody } from "./component";
 import { usePendingEntities } from "$hooks";
 import { Redirect } from "$components/Redirect";
 import { RoutesBuilder } from "$models/RoutesBuilder";
@@ -10,5 +10,5 @@ export const ListContainer: FunctionComponent = () => {
   if (response.loading) return <Fragment/>;
   if (response.error) return <Redirect to={RoutesBuilder.public.internalServerError()}/>;
 
-  return <List approvableEntities={response.data.getPendingEntities}/>;
+  return <ListBody approvableEntities={response.data.getPendingEntities}/>;
 };
